Spacemacs(layers): Add auto-completion variables
Set `nil` for return key behavior, because it interferes with inserting
newlines on `RET`.
Set `TAB` key behavior to `complete`, because auto-completion buffer can
be already cycled with `C-j` and `C-k`.
`jk` completion sequence is recommended by Spacemacs documentation and
allows even faster auto-completion, because these keys are located on
home row, as opposed to `TAB` which is quite harder to reach with pinky.